Add require_workspace_permission decorators to upload endpoints
Also refactor get_queryset for the nested viewsets into a new mixin.
