+++ multinet/api/views/common.py
@@ -1,10 +1,15 @@
 from typing import Dict, List
 
 from arango.cursor import Cursor
+from django.db.models import Model
+from django.http.response import Http404
+from django.shortcuts import get_object_or_404
 from drf_yasg import openapi
 from rest_framework.pagination import LimitOffsetPagination
 from rest_framework.request import Request
+from rest_framework_extensions.mixins import NestedViewSetMixin
 
+from multinet.api.models import Workspace
 from multinet.api.utils.arango import ArangoQuery
 
 
@@ -58,3 +63,24 @@ class ArangoPagination(LimitOffsetPagination):
         self.count = cur.statistics()['fullCount']
         self._set_post_query_params()
         return list(cur)
+
+
+class WorkspaceChildMixin(NestedViewSetMixin):
+    def get_queryset(self):
+        """
+        Get the queryset for workspace child enpoints endpoints. Check that the requeting user has
+        appropriate permissions for the associated workspace.
+        """
+
+        # prevent warning for schema generation incompatibility
+        if getattr(self, "swagger_fake_view", False):
+            return Model.objects.none()
+
+        child_objects = super().get_queryset()
+
+        parent_query_dict = self.get_parents_query_dict()
+        workspace = get_object_or_404(Workspace, name=parent_query_dict['workspace__name'])
+
+        if workspace.get_user_permission(self.request.user) is not None or workspace.public:
+            return child_objects
+        raise Http404
